+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Dev
+
+- Added multiple containers support(instead of scopes)
+- Added Resolver class
+- Container is refactored to use Resolver class
+
 ## 1.0.1
 
 - Added using functions and classes as keys